Edit button partialy complete

+++ 6-from-dusk-til-done/src/todoListItem.js
@@ -24,7 +24,18 @@ export default function todoListItem(todo) {
           .text(todo.text)
           .attr("for","checkbox"+todo.id)
       )
-      .addClass("todo col s10")
+      .addClass("todo col s9")
+  )
+  .append(
+    $(document.createElement("button"))
+      .addClass("btn-floating waves-effect waves-light teal edit")
+      .attr("id",`edit${todo.id}`)
+      .attr("type","button")
+      .append(
+        $(document.createElement("i"))
+          .addClass("material-icons")
+          .text("edit")
+      )
   )
   .append(
     $(document.createElement("button"))
@@ -37,8 +48,34 @@ export default function todoListItem(todo) {
       )
       .click((e) => {
         e.preventDefault();
-        deleteTodo([],todo.id);
+        if (confirm(`Are you sure you want to delete ${todo.text}?`)){deleteTodo([],todo.id)};
       })
+  )
+  .append(
+    $(document.createElement("div"))
+      .attr("id",`edit-todo-${todo.id}`)
+      .append(
+        $(document.createElement("input"))
+        .attr("type", "text")
+        .val(todo.text)
+        .attr("id",`edit-todo-text-${todo.id}`)
+      )
+      .append(
+        $(document.createElement("button"))
+          .attr("type","button")
+          .text("Save")
+          .click((e) => {
+            console.log("Saving todo");
+            const editedTodo = {
+              id: todo.id,
+              text: $(`#edit-todo-text-${todo.id}`).val(),
+              completed: $(`#checkbox${todo,id}`)
+            }
+            $(e.target).parent().slideUp();
+            $(``)
+          })
+      )
+      .hide();
   )
   .addClass("todoListItem row")
 }
